+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 dd mm yyyy
 
+- Change: Replaced the filesystem storage for the HTTP cache with a database storage. This fixes 
+  a problem with too many open files when `run-jobs.sh` is executed and there are many characters.
 - Change: Updated the ESI compatibility date to 2025-12-16.
 - Fix: Fixed a log message that said "error limit" instead of "rate limit".
 - Updated backend dependencies.